requirements.txt
Django==1.6.1
Pillow==2.3.0
South==0.8.4
boto==2.23.0
dj-database-url==0.2.2
dj-static==0.0.5
django-attachments==0.3.1
django-crispy-forms==1.4.0
django-registration==1.0
django-s3-collectstatic==1.0
django-storages==1.1.8
django-toolbelt==0.0.1
django-xadmin==0.4.2
easy-thumbnails==1.4
gunicorn==18.0
psycopg2==2.5.2
six==1.5.2
static==0.4
wsgiref==0.1.2
